requirements.txt
PyQt6
pyqt6-tools
mysql-connector-python